Fixing up based on refactor
